added script to compute ess standalone
Better solution than running a notebook for long periods of time.
New approach: compute values and store them as json. Load those values
in the notebook and visualize and explain them there.
This makes it easier to use the notebooks directly in writing and makes
them run faster. Furthermore, once the data is there it can be used
in different places!