svgd back in; with notebook to showcase
Stein Variational Gradient Descent was still missing in this repository;
this commit adds it back in.
This is not yet perfectly integrated into our framework, in particular
with respect to our BNN. The following desiderata are not yet met:
    1. SVGD should not take a single tf.Variable with all particles as
       input, but rather a list of particles - each a single
       tf.Variable. We might solve this issue by using tf.pack (or
       something similar) to obtain a single tensor from that list.
    2. BNN needs to add some code to support SVGD. In particular, instead of
       maintaining one network obtained by callable `get_net` it needs
       to maintain a network per SVGD particle.

After fixing these open issues, SVGD can be added to
`sampling.Sampler.get_sampler` and we should be able to fit BNNs using
SVGD, without any overhead for the user and without changing existing
code for any of the other samplers.