tests for seeding in samplers. fixed seed error
Relativistic SGHMC did not use proper seeding for adaptive rejection
sampling. This was fixed in ARSpy 0.4 which was just released. This
commit adapts relativistic sghmc to use these new seeding facilities.

Furthermore, we provide a (fast-running) testsuite for all supported
sampling methods that ensures that fixing an integer seed makes samples
and costs completely reproducible. This test suite will run on travis,
since reproducibility this is pretty important when experimenting with
any of the samplers.