changed interface for svgd
svgd interface is now closer to how it needs to be in order to use it
with BNN. it takes a list of `tensorflow.Variable` objects as input,
instead of a single tensorflow.Variable. This interface already allows
us to comfortably use it to sample from artificial functions (since it
strongly resembles the interface used in other `MCMCSampler`s.)

However, this does not yet fully resolve our issues with respect to
using it in BNNs. Using SVGD in BNNs requires maintaining multiple
network guesses at the same time. Ideally, one would specify multiple
such guesses using BayesianNeuralNetwork.get_net(). This gives a big
list of tensorflow.Variable objects. Then, we might group them
according to variable namespaces (into list of list) to get individual
guesses. This would then be passed to svgd, which computes costs for all
samples and averages them in each step to produce a single sample which
is returned to our BNN.
We have to see if this approach is computationally tractable, but it nicely
shields users from dealing with the intricacies of svgd vs MCMC methods.