provide nutspaper heuristic to initialize stepsize
This commit provides a integrated heuristic in stepsize_schedules that
can be used in conjunction with any stepsize schedule and will perform
a couple of leapfrog steps to determine a sound initial value for
stepsize epsilon on the given benchmark function.

Some refactoring is still necessary: integration into sgld sampler is
still not implemented and the interface is not super comfortable yet.
These changes will still be implemented on this current branch. After
this, we will write some more tests that check that this feature does
not regress performance of our samplers (also when used with BNN).
After that, we may merge this feature branch back into development/
branch.