adding functionality (and tests) to reset samplers
Resetting a sampler means resetting all its parameters to their initial
value. Currently we have a hack that simply resets *all* global
variables. It would be preferable to bookmark relevant parameters for
the current sampler and reset those instead.
NOTE: This does not allow full chain reproducibility for a sampler
(which cannot be achieved at all at the moment).
Tensorflow does not allow resetting its internal random stream,
so even with seeds it is not possible to get fork-safety in tensorflow.