initializing branch: towards keras
The goal of this branch is to use keras as our main interface to write
bnns (and samplers). All samplers are implemented as a
`keras.Optimizer` and first tested with our BNN before considered
ready. Using keras makes a number of things easier: integrating new
error metrics (a lot of preimplemented one can just be plugged-in),
unifying our interfaces (they already did that!), obtaining datasets
(it ships with cifar10, mnist and boston housing) etc.
We will still provide stand-alone samplers, which should be fairly easy
to do. Given a `keras.Optimizer` we can automatically generate a
sampler that has a similar interface as our old samplers did (in
particular `next` is still as feasible as it was.). Code to perform
this automatic generation will follow soon.
For now, our job is to reimplement old functionality in such a way that
we can train BNNs with all previously available samplers and start
integrating new ideas.